// src/types/enums/PathsRoute.ts
export enum PathsRoute {
  Home = "/",
  Cars = "/cars",
  Fuel = "/fuel",
  Service = "/service",
  Insurance = "/insurance",
  TechnicalInspection = "/technical-inspection",
  Profile = "/profile",
  Statistics = "/statistics",
  Blog = "/Blog",
  Contacts = "/contacts",
  AboutUs = "/about-us",
}